Renamed SlideTemplate1 to DefaultSlideTemplate
