+++ config/index.js
@@ -12,7 +12,7 @@ function getBrowserCommand () {
   // get the operating system version that we are running on
   const platform = process.platform
   if (platform === 'darwin') {// if we are running on mac os
-    return 'Applications/Google\ Chrome.app/Contents/MacOS/Google\ Chrome'
+    return '/Applications/Google\ Chrome.app/Contents/MacOS/Google\ Chrome'
   } else if (platform === 'win32'){ // if we are running on windows
     return 'chrome'
   } else if (platform === 'linux') { // if we are running on linux
