+++ src/components/Templates/DefaultSlideTemplate.vue
